Cleaned up Resources Section and added text inputs for title/url fields
