+++ src/edit-group/EditGroupPropTypes.js
@@ -1,0 +1,14 @@
+import PropTypes from 'prop-types';
+
+// eslint-disable-next-line import/prefer-default-export
+export const ResourceShape = PropTypes.shape({
+  id: PropTypes.string,
+  title: PropTypes.string,
+  action: PropTypes.string,
+  relatedNode: PropTypes.shape({
+    __typename: PropTypes.string,
+    id: PropTypes.string,
+    title: PropTypes.string,
+    url: PropTypes.string,
+  }),
+});
